fastapi_app/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.0
asyncpg==0.30.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.10
packaging==24.1
pendulum==3.0.0
pluggy==1.5.0
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
redis==5.2.0
rich==13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
time-machine==2.16.0
tomli==2.0.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
gunicorn
